Add support for BrokerSelectTag
Select a broker to use based on the broker having a specific
tag. (e.g. tagging brokers with which datacenters they serve
provides a straight-forward method for ops/dev to use when
new checks are created via this module by setting the
BrokerSelectTag accordingly - only brokers with that tag
would be considered valid.)