Change `Value` to be of type `interface{}`.
When the metric type for a ruleset is of type `text`, the value will
come back as a string.  When the metric type for a ruleset is `numeric`,
however, `value` will be sent as an integer or float64.  The docs say
this should be a string, however it appears as though this attribute is
polymorphic.  For the time being, use `interface{}` until this gets
sorted out at the API (if ever).