Mechanically change and "fix" the types returned from the API.
Burried in the documentation of `encoding/json`'s `Marshal()`
docs (https://golang.org/pkg/encoding/json/#Marshal) there's a beautiful
little nugget referencing a useful `json` encoding option: `string`.

> The "string" option signals that a field is stored as JSON inside a
> JSON-encoded string. It applies only to fields of string, floating
> point, integer, or boolean types. This extra level of encoding is
> sometimes used when communicating with JavaScript programs

The mechanical fix was to use this everywhere that a BUG was listed and
change the type to a `*int`.  Attributes with a `*int` are traversed
correctly and signal a missing value without relying on the zero value
of parameters.

While I can't guarantee that this is 100% correct PR and that all of
these attributes do in fact allow `omitempty` to be set as was the case
previously (see exhibit `Stack` in struct `GraphDatapoint`), it does
appear that we are at least as well off as we were, and are likely
incrementally better despite not being certain of 100% correctness.